+++ src/Phases/battlePhase.cpp
@@ -18,7 +18,7 @@ namespace // namespace_start
 #define ATTACK 0
 #define DEFEND 1
 
-size_t ptsDiff;
+uint16_t ptsDiff;
   
 PlayerPtr player;
 PlayerPtr enemy;
@@ -34,8 +34,8 @@ ProvincePtr chooseProvince(PlayerPtr player);
 
 int chooseAction(void);
 
-size_t calcTotalATK(ArmyPtr battleArmy);
-size_t calcTotalDEF(ArmyPtr battleArmy);
+uint16_t calcTotalATK(ArmyPtr battleArmy);
+uint16_t calcTotalDEF(ArmyPtr battleArmy);
 
 void chooseArmy(PlayerPtr player, ArmyPtr battleArmy);
 void provinceDestroyed(ProvincePtr prov);
@@ -118,7 +118,7 @@ ProvincePtr chooseProvince(PlayerPtr player)
   player->printProvinces();
 
   std::string answer;
-  size_t prov;
+  uint16_t prov;
 
   while (true)
   {
@@ -135,7 +135,7 @@ Your input should be an integer in range [1," << player->getProvincesNum()
     cout << "Wrong input given. Please, try again." << endl;
   }
   
-  size_t counter = 0;
+  uint16_t counter = 0;
   for (auto i : *(player->getProvinces()))
   {
     if (i->checkBroken() == true) continue;
@@ -172,9 +172,9 @@ appearance, to recruit the Personality for the Battle!" << endl;
 
 /* ========================================================================= */
 
-size_t calcTotalATK(ArmyPtr battleArmy)
+uint16_t calcTotalATK(ArmyPtr battleArmy)
 {
-  size_t totalATK = 0;
+  uint16_t totalATK = 0;
 
   for (std::shared_ptr< Personality > i : *battleArmy) // sum ATK of all personalities -> sum ATK of all their flollowers + items
   {
@@ -192,9 +192,9 @@ size_t calcTotalATK(ArmyPtr battleArmy)
 
 /* ========================================================================= */
 
-size_t calcTotalDEF(ArmyPtr battleArmy)
+uint16_t calcTotalDEF(ArmyPtr battleArmy)
 {
-  size_t totalDEF = 0;
+  uint16_t totalDEF = 0;
 
   for (auto i : *battleArmy) // sum DEF of all personalities -> sum DEF of all their flollowers + items
   {
@@ -292,7 +292,7 @@ void attackerWins()
   for (auto i : *defArmy) i->die(); // todo: cleanup dead from their lists
   cout << "Defender's army is destroyed." << endl;
 
-  //size_t ptsDiff = attPoints - defPoints;
+  //uint16_t ptsDiff = attPoints - defPoints;
 
   cout << "Attacker's army has experienced heavy casualties." << endl; 
   verifyCasualties(attArmy, ATTACK);
@@ -307,7 +307,7 @@ void defenderWins()
   for (auto i : *attArmy) i->die(); // todo: cleanup dead from their lists
   cout << "Attacker's army is destroyed." << endl;
 
-  // size_t ptsDiff = attPoints - defPoints;
+  // uint16_t ptsDiff = attPoints - defPoints;
 
   cout << "Defender's army has experienced heavy casualties." << endl; 
   verifyCasualties(attArmy, DEFEND);
@@ -316,8 +316,8 @@ void defenderWins()
 
 void battle(ProvincePtr prov)
 {
-  size_t attPoints = calcTotalATK(attArmy);
-  size_t defPoints = calcTotalDEF(defArmy);// + enemy->getStrongHold()->getInitDEF(); // todo: verify @lists
+  uint16_t attPoints = calcTotalATK(attArmy);
+  uint16_t defPoints = calcTotalDEF(defArmy);// + enemy->getStrongHold()->getInitDEF(); // todo: verify @lists
 
   cout << "Attacker's ARMY strength is: " << attPoints << endl;
   cout << "Defender's ARMY strength is: " << defPoints << endl;
